Fixed warnings.
git-svn-id: http://sproxel.googlecode.com/svn/trunk@16 f3b4058f-8ea4-8f77-95d1-b9d5bc24aafc

+++ GameVoxelGrid.h
@@ -116,7 +116,7 @@ public:
         // Return only what's needed
         std::vector<Imath::V3i> orderedCells;
         orderedCells.resize(sortList.size());
-        for (int i = 0; i < sortList.size(); i++)
+        for (size_t i = 0; i < sortList.size(); i++)
             orderedCells[i] = sortList[i].index;
         return orderedCells;
     }
